Extract useDeviceEnrolment so Workers can show a compact code box
Code entry, status polling and approve/deny lived inside AddMachinePanel,
which meant the only place to type an enrolment code was inside the fully
expanded panel. Someone setting up a headless machine has the code in hand
already and just wants somewhere to put it.

Lifting that state into a hook lets Workers.tsx render a compact code box
next to the collapsed "Add a machine" summary while the full panel
underneath binds to the very same state, so the two can't drift. The
collapsed box stays interactive, with stopPropagation keeping clicks and
typing from toggling the <details>.

The panel also now opens itself once a typed code resolves to something
needing action -- pending confirmation, or an already-approved message --
so a code entered while collapsed doesn't leave its own result hidden. It
starts open when the account has no machines yet.
