+++ client/src/components/AddMachinePanel.tsx
@@ -1,6 +1,5 @@
-import { useEffect, useMemo, useState } from "react";
-import { api } from "../api/client";
-import type { DeviceStatusResponse } from "../types";
+import { useMemo, useState } from "react";
+import { useDeviceEnrolment } from "../api/useDeviceEnrolment";
 import {
   CopyCommandButton,
   PowerShellNotice,
@@ -12,121 +11,35 @@ import {
 import { IconServer, IconInfo, IconCheck } from "./icons";
 import { formatRelativeTime } from "../utils";
 
-// Matches server/src/session.ts's generateUserCode (4 chars, a dash, 4 more,
-// drawn from Crockford base32 minus ambiguous characters) -- used both to
-// know when the field is "full enough to poll" and to reject obviously
-// unfinished input before it ever reaches the server.
-const CODE_RE = /^[A-Z0-9]{4}-[A-Z0-9]{4}$/;
-
-function normalizeCodeInput(raw: string): string {
-  const upper = raw
-    .toUpperCase()
-    .replace(/[^A-Z0-9]/g, "")
-    .slice(0, 8);
-  return upper.length > 4 ? `${upper.slice(0, 4)}-${upper.slice(4)}` : upper;
-}
-
-const POLL_INTERVAL_MS = 2000;
-
 // MULTIUSER_PLAN.md §3.1's "Add machine" screen, extracted from pages/
 // Device.tsx so it can be embedded directly on the Workers page (top of the
 // page) AND still rendered by the bookmarkable /device route -- it's the
 // SAME component either way, not two separate screens. Assumes a logged-in
 // session with AUTH_ENABLED (the /device route checks that itself before
 // rendering this).
-export function AddMachinePanel() {
+//
+// Enrolment-code state (Workers.tsx also shows a compact code box next to
+// the collapsed "Add a machine" summary, wired to this same state) can be
+// passed in via `enrolment`; when omitted (the /device route's standalone
+// use) this component creates its own.
+export function AddMachinePanel({
+  enrolment,
+}: {
+  enrolment?: ReturnType<typeof useDeviceEnrolment>;
+}) {
   // Pre-selected from the browser's own OS (§1 of the Settings/Device
   // rework) so the right setup command is already showing on first paint --
   // still just the tab's initial value, so a click on WIN/MACOS/LINUX above
   // freely overrides it same as before.
   const [os, setOs] = useState<SetupOS>(() => detectOS());
-  const [codeInput, setCodeInput] = useState("");
-  const [status, setStatus] = useState<DeviceStatusResponse | null>(null);
-  const [busy, setBusy] = useState(false);
-  const [error, setError] = useState<string | null>(null);
-  const [approved, setApproved] = useState<{ hostname: string; merged: boolean } | null>(null);
+  const ownEnrolment = useDeviceEnrolment();
+  const { codeInput, handleCodeChange, validCode, status, busy, error, approved, handleApprove, handleDeny } =
+    enrolment ?? ownEnrolment;
 
   // This page's origin IS the server's public URL (see
   // WorkerCard.tsx's buildSetupScenarios doc comment) -- stable for the
   // life of the page, so computed once rather than on every render.
   const setupScenarios = useMemo(() => buildSetupScenarios(window.location.origin), []);
-
-  const validCode = CODE_RE.test(codeInput);
-
-  // Polls GET /api/device/status while a full-length code is present (§3.1
-  // step 4) -- stops as soon as it resolves to "approved" (whether from this
-  // tab's own Approve click below, or a second browser tab/device that
-  // approved it first) or the input changes to something no longer a full
-  // code.
-  useEffect(() => {
-    if (!validCode) return;
-    let cancelled = false;
-    let timer: ReturnType<typeof setInterval> | null = null;
-    async function poll() {
-      try {
-        const s = await api.getDeviceStatus(codeInput);
-        if (cancelled) return;
-        setStatus(s);
-        setError(null);
-        if (s.state === "approved" && timer) {
-          clearInterval(timer);
-          timer = null;
-        }
-      } catch (err) {
-        if (!cancelled) setError(err instanceof Error ? err.message : String(err));
-      }
-    }
-    void poll();
-    timer = setInterval(poll, POLL_INTERVAL_MS);
-    return () => {
-      cancelled = true;
-      if (timer) clearInterval(timer);
-    };
-  }, [validCode, codeInput]);
-
-  function handleCodeChange(raw: string) {
-    setCodeInput(normalizeCodeInput(raw));
-    setStatus(null);
-    setError(null);
-  }
-
-  async function handleApprove(mergeInto?: string) {
-    setBusy(true);
-    setError(null);
-    try {
-      // confirm_duplicate reflects what the last status poll already showed
-      // -- see server/src/routes/device.ts's POST /api/device/approve and
-      // workerRepo.findPossibleDuplicate's doc comment for why this exists:
-      // deleting a worker's install folder wipes its persisted machine_id,
-      // so re-running setup looks like a brand-new machine to the server,
-      // and without this check would silently create an indistinguishable
-      // duplicate of a machine the user already has. mergeInto instead asks
-      // the server to re-attach this connection to that existing machine.
-      const confirmDuplicate = !mergeInto && status?.state === "pending" && status.possibleDuplicate != null;
-      const res = await api.approveDevice(codeInput, confirmDuplicate, mergeInto);
-      if (!res.ok) {
-        // Race: the duplicate was only detected server-side just now (this
-        // poll hadn't caught up yet) -- the next poll tick will pick up
-        // possibleDuplicate and relabel the buttons; nothing was approved.
-        setError('This machine looks like one you already have -- pick "Merge" or "Add as new machine" once it appears below.');
-        return;
-      }
-      setApproved({ hostname: res.machine.hostname ?? "This machine", merged: res.merged === true });
-    } catch (err) {
-      setError(err instanceof Error ? err.message : String(err));
-    } finally {
-      setBusy(false);
-    }
-  }
-
-  // No dedicated deny endpoint exists (or needs to) -- a code nobody approves
-  // just expires on its own in 15 minutes (server/src/routes/device.ts's
-  // ENROLMENT_TTL_MS). This just walks the user away from the confirm card.
-  function handleDeny() {
-    setCodeInput("");
-    setStatus(null);
-    setError(null);
-  }
 
   return (
     <div className="max-w-2xl">
@@ -191,20 +104,28 @@ export function AddMachinePanel() {
         </div>
       ) : (
         <div className="mt-6">
-          <label htmlFor="device-code" className="text-sm font-semibold text-fg">
-            Have a code?
-          </label>
-          <div>
-            <input
-              id="device-code"
-              type="text"
-              value={codeInput}
-              onChange={(e) => handleCodeChange(e.target.value)}
-              placeholder="ABCD-EFGH"
-              maxLength={9}
-              className="mt-1.5 w-40 rounded-lg border border-border bg-surface px-3 py-2 text-center font-mono text-sm uppercase tracking-widest text-fg outline-none focus:border-accent/50"
-            />
-          </div>
+          {/* When embedded on the Workers page (`enrolment` passed in), the
+              same code box already sits next to the "Add a machine" summary
+              above -- showing it again here would just be a second box bound
+              to the same state. Standalone (/device route) still needs it. */}
+          {!enrolment && (
+            <>
+              <label htmlFor="device-code" className="text-sm font-semibold text-fg">
+                Have a code?
+              </label>
+              <div>
+                <input
+                  id="device-code"
+                  type="text"
+                  value={codeInput}
+                  onChange={(e) => handleCodeChange(e.target.value)}
+                  placeholder="ABCD-EFGH"
+                  maxLength={9}
+                  className="mt-1.5 w-40 rounded-lg border border-border bg-surface px-3 py-2 text-center font-mono text-sm uppercase tracking-widest text-fg outline-none focus:border-accent/50"
+                />
+              </div>
+            </>
+          )}
 
           {error && <p className="mt-2 text-sm text-danger">{error}</p>}
